Moved and renamed generic styled component from itempage to App
