+++ src/components/Cart.js
@@ -1,8 +1,8 @@
 // import styled from "styled-components";
-import React from "react";
+import React from 'react'
 
 const Cart = () => {
-  return <div></div>;
-};
+  return <div></div>
+}
 
-export default Cart;
+export default Cart
