+++ src/components/ItemList.js
@@ -6,25 +6,6 @@ import { fetchItemsList } from '../features/itemListSlice'
 import Container from './Container'
 import ItemCard from './ItemCard'
 //#region  styled components
-const StyledItemsPage = styled('div')`
-  color: white;
-  width: 95vw;
-  margin: 0 auto;
-  padding: 2vh 5vh 0vh 5vh;
-`
-const StyledLogo = styled('div')`
-  display: block;
-  width: 5vw;
-  height: 5vw;
-  border: 1px solid white;
-  margin: 0 auto;
-`
-
-const StyledAppSeparator = styled('div')`
-  border-bottom: 1px solid white;
-  width: 100%;
-  margin: 5vh auto;
-`
 
 //#endregion
 const ItemList = () => {
@@ -37,9 +18,7 @@ const ItemList = () => {
   const inventory = useSelector(state => state.itemsData.inventory)
 
   return (
-    <StyledItemsPage>
-      <StyledLogo />
-      <StyledAppSeparator />
+    <>
       {inventory && (
         <Container>
           {inventory.map(props => {
@@ -48,7 +27,7 @@ const ItemList = () => {
         </Container>
       )}
       {/* <StyledButton /> */}
-    </StyledItemsPage>
+    </>
   )
 }
